requirements.txt
fastapi==0.104.1
imbalanced_learn==0.10.1
joblib==1.3.2
matplotlib==3.8.0
numpy==1.26.1
onnxruntime_gpu==1.16.1
pandas==2.1.2
pydantic==2.4.2
python_docx==1.1.0
Requests==2.31.0
scikit_learn==1.3.2
selenium==4.15.2
tensorflow==2.14.0
tensorflow_intel==2.14.0
torch==2.1.0
tqdm==4.66.1
wordcloud==1.9.2
